requirements.txt
# Indispensable libraries
Pillow==7.0.0 
pytube3==9.6.4
urllib3==1.25.8